Make it work for pocket accounts / custom product.
Fixed possible nullable properties.
