+++ accounts-journey/src/main/kotlin/com/backbase/accounts_journey/common/Mapper.kt
@@ -1,5 +1,6 @@
 package com.backbase.accounts_journey.common
 
+@Suppress("TooGenericExceptionCaught", "SwallowedException")
 inline fun <T, R> T.mapCatching(block: T.() -> R): R? {
     return try {
         block()
